+++ CHANGELOG.md
@@ -23,6 +23,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Add template function `atan_polyApprox` which is reduced to `atan_polyApprox_deg7` or `atan_polyApprox_deg9` according to its template parameter.
   - Add template function `atan2_polyApprox` whose polynomial degree is determined at compile-time
 - Added `addSqMat` function.
+- Made the following small functions target of aggressive inline-expansion:
+  - `LinAlg::`
+    - `isEqualMat`, `complexMat`, `conjugateMat`, `addMat`, `scaleMat`
+- Forcibly inline-expand the following small functions:
+  - `LinAlg::`
+    - `isEqualVec`, `complexVec`, `conjugateVec`, `addVec`, `scaleVec`, `l2Norm`
 
 ## [0.4.0] - 2021/8/27